backend/pyproject.toml
[tool.poetry]
name = "voyeur"
version = "0.1.0"
description = "A Django project for voyeur service"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
Django = "^5.1.5"
djangorestframework = "^3.14.0"
channels = "^4.0.0"
stomp-py = "^7.0.0"
pymongo = "^4.0.0"
django-cors-headers = "^3.13.0"
flask = "^2.0.0" 
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
